\name{SEL-internal}
\title{SEL internal functions}
\alias{get.Ab}
\alias{get.PmatBern}
\alias{get.PmatBS}
\alias{int.funct}
\alias{itgrl}
\alias{prod.Bspline}
\alias{zero}
\alias{print.SEL}
\alias{print.summary.SEL}
\alias{summary.SEL}
\alias{coef.SEL}
\description{
  These functions are not meant to be called by the user.
}

\keyword{ internal }
